Refatoração Backend
